52: Moved Notification component to index.js
